requirements.txt
numpy>=1.26
pandas>=2.1
pyyaml>=6.0
pydantic>=2.6
pydantic-settings>=2.2
scikit-learn>=1.4
torch>=2.2
tqdm>=4.66
mlflow>=2.10
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
httpx>=0.27
